Add LoginForm component for user login
